# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cache/translations": {
      "get": {
        "operationId": "CacheTranslationsController_getTranslations",
        "parameters": [
          {
            "name": "searchType",
            "required": false,
            "in": "query",
            "description": "검색 기준",
            "schema": {
              "example": "source",
              "type": "string",
              "enum": [
                "source",
                "target",
                "fileName",
                "filePath",
                "date",
                "cacheTag"
              ]
            }
          },
          {
            "name": "searchValue",
            "required": false,
            "in": "query",
            "description": "검색어",
            "schema": {
              "example": "예제",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "검색 시작일 (YYYY-MM-DD)",
            "schema": {
              "example": "2024-11-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "검색 종료일 (YYYY-MM-DD)",
            "schema": {
              "example": "2024-12-01",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "요청할 페이지 번호 (1부터 시작)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "itemsPerPage",
            "required": false,
            "in": "query",
            "description": "페이지당 항목 수",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "번역 캐시 목록을 조회합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCacheTranslationsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "cache-translations"
        ]
      },
      "delete": {
        "operationId": "CacheTranslationsController_deleteTranslations",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteCacheTranslationsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "선택된 번역을 삭제하거나 검색 조건에 맞는 모든 번역을 삭제합니다. `translationIds` 또는 `searchParams` 중 하나는 필수입니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCacheTranslationsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "cache-translations"
        ]
      }
    },
    "/cache/translations/{translationId}/history": {
      "get": {
        "operationId": "CacheTranslationsController_getTranslationHistory",
        "parameters": [
          {
            "name": "translationId",
            "required": true,
            "in": "path",
            "description": "대상 번역 ID",
            "schema": {
              "example": 12,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "선택한 번역의 이력을 조회합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTranslationHistoryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "cache-translations"
        ]
      }
    },
    "/cache/translations/{translationId}": {
      "patch": {
        "operationId": "CacheTranslationsController_updateTranslation",
        "parameters": [
          {
            "name": "translationId",
            "required": true,
            "in": "path",
            "description": "대상 번역 ID",
            "schema": {
              "example": 12,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTranslationBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "번역 결과를 수정합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateTranslationResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "cache-translations"
        ]
      }
    },
    "/cache/translations/{translationId}/cache-tag": {
      "patch": {
        "operationId": "CacheTranslationsController_updateTranslationCacheTag",
        "parameters": [
          {
            "name": "translationId",
            "required": true,
            "in": "path",
            "description": "대상 번역 ID",
            "schema": {
              "example": 12,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTranslationCacheTagBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "번역의 캐시 태그를 변경합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateTranslationCacheTagResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "cache-translations"
        ]
      }
    },
    "/cache/translations/export": {
      "get": {
        "operationId": "CacheTranslationsController_exportTranslations",
        "parameters": [
          {
            "name": "searchType",
            "required": false,
            "in": "query",
            "description": "검색 기준",
            "schema": {
              "example": "source",
              "type": "string",
              "enum": [
                "source",
                "target",
                "fileName",
                "filePath",
                "date",
                "cacheTag"
              ]
            }
          },
          {
            "name": "searchValue",
            "required": false,
            "in": "query",
            "description": "검색어",
            "schema": {
              "example": "예제",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "검색 시작일 (YYYY-MM-DD)",
            "schema": {
              "example": "2024-11-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "검색 종료일 (YYYY-MM-DD)",
            "schema": {
              "example": "2024-12-01",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "검색 조건에 맞는 번역을 내보냅니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportTranslationsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "cache-translations"
        ]
      }
    },
    "/cache/translations/import": {
      "post": {
        "operationId": "CacheTranslationsController_importTranslations",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportTranslationsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "번역을 가져옵니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportTranslationsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "cache-translations"
        ]
      }
    },
    "/cache/tags": {
      "get": {
        "operationId": "CacheTagsController_getCacheTags",
        "parameters": [
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "태그 이름 검색 키워드",
            "schema": {
              "example": "회의",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "정렬 기준",
            "schema": {
              "type": "string",
              "enum": [
                "lastUsedAt",
                "name",
                "createdAt"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "정렬 순서",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "캐시 태그 목록을 조회합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetCacheTagsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "cache-tags"
        ]
      }
    },
    "/cache/tags/{id}": {
      "delete": {
        "operationId": "CacheTagsController_deleteCacheTag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "삭제할 캐시 태그 ID",
            "schema": {
              "example": 9,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteCacheTagBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "캐시 태그를 삭제합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCacheTagResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "cache-tags"
        ]
      }
    },
...
  },
  "info": {
    "title": "FormattedDocs API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CacheTranslationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "번역 캐시 ID",
            "example": 42
          },
          "source": {
            "type": "string",
            "description": "원본 텍스트",
            "example": "Hello"
          },
          "target": {
            "type": "string",
            "description": "번역된 텍스트",
            "example": "안녕하세요"
          },
          "createdAt": {
            "type": "string",
            "description": "항목이 생성된 시각",
            "example": "2024-12-01T12:34:56.000Z"
          },
          "lastAccessedAt": {
            "type": "string",
            "description": "마지막으로 접근된 시각",
            "example": "2024-12-10T09:20:00.000Z"
          },
          "cacheTag": {
            "type": "string",
            "description": "연결된 캐시 태그 이름",
            "example": "일반"
          },
          "cacheTagId": {
            "type": "number",
            "description": "연결된 캐시 태그 ID, 없을 경우 null",
            "example": 3,
            "nullable": true
          },
          "fileName": {
            "type": "string",
            "description": "연관된 파일 이름",
            "example": "document.txt",
            "nullable": true
          },
          "filePath": {
            "type": "string",
            "description": "연관된 파일 경로",
            "example": "/Users/username/Documents/document.txt",
            "nullable": true
          }
        },
        "required": [
          "id",
          "source",
          "target",
          "createdAt",
          "lastAccessedAt",
          "cacheTag"
        ]
      },
      "GetCacheTranslationsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "translations": {
            "description": "조회된 번역 목록",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CacheTranslationDto"
            }
          },
          "totalItems": {
            "type": "number",
            "description": "전체 항목 수",
            "example": 120
          }
        },
        "required": [
          "success",
          "translations",
          "totalItems"
        ]
      },
      "TranslationHistoryDto": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "description": "원본 텍스트",
            "example": "Hello"
          },
          "target": {
            "type": "string",
            "description": "번역된 텍스트",
            "example": "안녕하세요"
          },
          "success": {
            "type": "boolean",
            "description": "번역 성공 여부",
            "example": true
          },
          "error": {
            "type": "string",
            "description": "번역 실패 시 오류 메시지",
            "example": "Rate limit exceeded",
            "nullable": true
          },
          "model": {
            "type": "string",
            "description": "번역 요청에 사용된 모델 이름",
            "example": "gpt-4o-mini"
          },
          "createdAt": {
            "type": "string",
            "description": "이력이 기록된 시각",
            "example": "2024-12-10T12:00:00.000Z"
          },
          "cacheTag": {
            "type": "string",
            "description": "번역 시 사용된 캐시 태그 이름",
            "example": "일반"
          }
        },
        "required": [
          "source",
          "target",
          "success",
          "model",
          "createdAt",
          "cacheTag"
        ]
      },
      "GetTranslationHistoryResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "translationHistory": {
            "description": "선택한 번역의 이력 목록",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranslationHistoryDto"
            }
          }
        },
        "required": [
          "success",
          "translationHistory"
        ]
      },
      "UpdateTranslationBodyDto": {
        "type": "object",
        "properties": {
          "target": {
            "type": "string",
            "description": "수정할 번역 텍스트",
            "example": "안녕하세요!"
          }
        },
        "required": [
          "target"
        ]
      },
      "UpdateTranslationResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          }
        },
        "required": [
          "success"
        ]
      },
      "UpdateTranslationCacheTagBodyDto": {
        "type": "object",
        "properties": {
          "cacheTagId": {
            "type": "number",
            "description": "변경할 캐시 태그 ID",
            "example": 5
          }
        },
        "required": [
          "cacheTagId"
        ]
      },
      "UpdateTranslationCacheTagResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          }
        },
        "required": [
          "success"
        ]
      },
      "CacheTranslationsSearchParamsDto": {
        "type": "object",
        "properties": {
          "searchType": {
            "type": "string",
            "description": "검색 기준",
            "enum": [
              "source",
              "target",
              "fileName",
              "filePath",
              "date",
              "cacheTag"
            ],
            "example": "source"
          },
          "searchValue": {
            "type": "string",
            "description": "검색어",
            "example": "예제"
          },
          "startDate": {
            "type": "string",
            "description": "검색 시작일 (YYYY-MM-DD)",
            "example": "2024-11-01"
          },
          "endDate": {
            "type": "string",
            "description": "검색 종료일 (YYYY-MM-DD)",
            "example": "2024-12-01"
          }
        }
      },
      "DeleteCacheTranslationsRequestDto": {
        "type": "object",
        "properties": {
          "translationIds": {
            "description": "삭제할 번역 ID 목록",
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "searchParams": {
            "description": "검색 조건에 해당하는 모든 번역 삭제 시 사용되는 조건",
            "allOf": [
              {
                "$ref": "#/components/schemas/CacheTranslationsSearchParamsDto"
              }
            ]
          }
        }
      },
      "DeleteCacheTranslationsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          }
        },
        "required": [
          "success"
        ]
      },
      "TranslationExportImportDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "번역 ID",
            "example": 101
          },
          "source": {
            "type": "string",
            "description": "원본 텍스트",
            "example": "How are you?"
          },
          "target": {
            "type": "string",
            "description": "번역 텍스트",
            "example": "잘 지내?"
          },
          "cacheTag": {
            "type": "string",
            "description": "연결된 캐시 태그",
            "example": "일반"
          }
        },
        "required": [
          "id",
          "source",
          "target",
          "cacheTag"
        ]
      },
      "ExportTranslationsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "translations": {
            "description": "내보낸 번역 목록",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranslationExportImportDto"
            }
          }
        },
        "required": [
          "success",
          "translations"
        ]
      },
...
      "ImportTranslationsRequestDto": {
        "type": "object",
        "properties": {
          "translations": {
            "description": "가져올 번역 목록",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportTranslationItemDto"
            }
          }
        },
        "required": [
          "translations"
        ]
      },
      "ImportTranslationsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "updatedCount": {
            "type": "number",
            "description": "업데이트되거나 생성된 번역 수",
            "example": 25
          }
        },
        "required": [
          "success",
          "updatedCount"
        ]
      },
...
      "GetCacheTagsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "cacheTags": {
            "description": "캐시 태그 목록",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CacheTagSummaryDto"
            }
          }
        },
        "required": [
          "success",
          "cacheTags"
        ]
      },
      "DeleteCacheTagBodyDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "description": "삭제 모드",
            "enum": [
              "strict",
              "cascade",
              "reassign"
            ],
            "default": "strict"
          },
          "targetTagId": {
            "type": "number",
            "description": "재할당 대상 캐시 태그 ID (mode가 reassign일 경우 필수)",
            "example": 3
          }
        }
      },
      "DeleteCacheTagResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          }
        },
        "required": [
          "success"
        ]
      },
...
    }
  }
}